Exposed licences in baseobject API
* api imports cleanup (circular imports) - created new module api_simple for simple serializers
* added licences to BaseObject serializer (inherited in BackOfficeAssetSerializer and DataCenterAsset serializer)
